mysite/requirements.txt
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.4.1
asttokens==2.0.5
atomicwrites==1.4.0
attrs==21.4.0
autopep8==1.6.0
backcall==0.2.0
black==21.12b0
bleach==4.1.0
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.10
click==8.0.3
colorama==0.4.4
cryptography==36.0.1
cycler==0.11.0
debugpy==1.5.1
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.4
Django==4.0
django-bootstrap-datepicker-plus==4.0.0
django-crispy-forms==1.13.0
django-environ==0.4.5
django-extensions==3.1.5
django-model-utils==4.2.0
django-pwa==1.0.10
entrypoints==0.3
executing==0.8.2
filelock==3.4.2
fonttools==4.28.2
idna==2.10
iniconfig==1.1.1
ipykernel==6.7.0
ipython==8.0.1
ipython-genutils==0.2.0
ipywidgets==7.6.5
jedi==0.18.1
Jinja2==3.0.3
jsonschema==4.4.0
jupyter-client==7.1.1
jupyter-core==4.9.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.2
kiwisolver==1.3.2
MarkupSafe==2.0.1
matplotlib==3.5.0
matplotlib-inline==0.1.3
mercadopago==2.0.6
mistune==0.8.4
mypy-extensions==0.4.3
nbclient==0.5.10
nbconvert==6.4.0
nbformat==5.1.3
nest-asyncio==1.5.4
notebook==6.4.7
numpy==1.22.2
oauthlib==3.1.1
packaging==21.3
pandocfilters==1.5.0
parso==0.8.3
pathspec==0.9.0
pickleshare==0.7.5
Pillow==8.4.0
platformdirs==2.4.1
pluggy==1.0.0
prometheus-client==0.12.0
prompt-toolkit==3.0.24
pure-eval==0.2.1
py==1.11.0
pycodestyle==2.8.0
pycparser==2.21
Pygments==2.11.2
PyJWT==2.3.0
pyOpenSSL==22.0.0
pyparsing==3.0.6
pyrsistent==0.18.1
pytest==7.0.1
pytest-mock==3.5.1
python-dateutil==2.8.2
python-dotenv==0.19.2
python-social-auth==0.3.6
python3-openid==3.2.0
pytz==2021.1
pywin32==303
pywinpty==1.1.6
pyzmq==22.3.0
requests==2.25.1
requests-oauthlib==1.3.0
Send2Trash==1.8.0
setuptools-scm==6.3.2
six==1.16.0
social-auth-app-django==5.0.0
social-auth-core==4.1.0
sqlparse==0.4.1
stack-data==0.1.4
terminado==0.12.1
testpath==0.5.0
toml==0.10.2
tomli==1.2.2
tornado==6.1
traitlets==5.1.1
typing_extensions==4.0.1
tzdata==2021.5
urllib3==1.26.8
virtualenv==20.12.1
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.0.3
whitenoise==6.0.0
widgetsnbextension==3.5.2
# windows-curses==2.3.0
gunicorn
psycopg2
